% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central.R
\docType{data}
\name{central}
\alias{central}
\title{Central pixel time series}
\format{A data frame with 8760 LST observations (rows, one year sampled hourly) and 1 variable (pixel):
\describe{
  \item{x}{pixel}
}}
\usage{
central
}
\description{
This dataset contains LST values in K from the central pixel
}
\keyword{datasets}
