% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Frame pixel time series}
\format{An object of class \code{data.frame} with 8760 rows and 4 columns.}
\usage{
data
}
\description{
This dataframe contains LST values in K from the four pixels of the frame
}
\keyword{datasets}
