% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_stl.R
\name{dec_stl}
\alias{dec_stl}
\title{Time series decomposition}
\usage{
dec_stl(series, n)
}
\arguments{
\item{series}{The time series to be decomposed, as a single column dataframe}

\item{n}{One or all of the components of the decomposition, or their combinations. Should be named "trend", "seasonal", "remainder", "trendseas", "trendrem", or "seasrem"}
}
\description{
The function performs Seasonal-Trend Decomposition by Loess, based on package stlplus (required). It plots and returns the component(s) defined by the user.
}
\author{
Effie Pavlidou
}
