\name{import}
\alias{import}
#'
#'@title
#'@description
#'@param n
#'@examples
#'
#'@author Effie Pavlidou
#'@export
#'
\title{import
}
\description{Function to input multiple time series. The series should be stored as 'f*.txt' (f1.txt, f2.txt etc).
}
\usage{
import(n)
}

\arguments{
  \item{n}{The number of series to be imported
}
}

\value{The function returns a dataframe where each column corresponds to one pixel's time series.
}
\author{Effie Pavlidou
}
