% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_command_translate.R
\name{tex_command_translate}
\alias{tex_command_translate}
\title{Replace latex index directives}
\usage{
tex_command_translate(command, arg1, arg2)
}
\description{
Replace latex index directives
}
