% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un_central.R
\docType{data}
\name{un_central}
\alias{un_central}
\title{Central pixel uncertainty time series}
\format{An object of class \code{data.frame} with 8760 rows and 1 columns.}
\usage{
un_central
}
\description{
This dataset contains uncertainties of LST values in K from the central pixel
}
\keyword{datasets}
