% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainties.R
\docType{data}
\name{uncertainties}
\alias{uncertainties}
\title{Frame pixel uncertainty time series}
\format{An object of class \code{data.frame} with 8760 rows and 4 columns.}
\usage{
uncertainties
}
\description{
This dataframe contains uncertainties of LST values in K from the four pixels of the frame
}
\keyword{datasets}
